pyproject.toml
# pyproject.toml
[project]
name = "edge-vlm-server"
version = "0.2.0"
description = "High-performance Edge VLM Inference Server with Vision-Language Model support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "requests>=2.30.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
vlm = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.25.0",
    "qwen-vl-utils>=0.0.8",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"